moved some treesitter code into ts file
